converted skeleton and product data in different small components
